Create querying mixin
